+++ src/codeSigning/upload.ts
@@ -14,7 +14,10 @@ import ClientConfiguration = WebIdentityCredentials.ClientConfiguration;
 export async function putObject(bucket: string, key: string, body: string): Promise<AWS.S3.PutObjectOutput> {
   return new Promise((resolve, reject) => {
     const agent = api.getAgentForUri('https://s3.amazonaws.com') as Agents;
-    const s3 = new AWS.S3({ httpsOptions: { agent: agent.https } } as ClientConfiguration);
+    const s3 = new AWS.S3({
+      httpOptions: { agent: agent.http },
+      httpsOptions: { agent: agent.https },
+    } as ClientConfiguration);
     s3.putObject({ Bucket: bucket, Key: key, Body: body }, (err, resp) => {
       if (err) reject(err);
       if (resp) resolve(resp);