+++ CHANGELOG.md
@@ -2,6 +2,25 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [1.13.5](https://github.com/salesforcecli/plugin-release-management/compare/v1.13.4...v1.13.5) (2021-06-28)
+
+### Bug Fixes
+
+- add registry param to sfdx-trust calls ([#215](https://github.com/salesforcecli/plugin-release-management/issues/215)) ([237e2f2](https://github.com/salesforcecli/plugin-release-management/commit/237e2f2e677a6833fb103e102ff1da8a768f33e1))
+
+### [1.13.4](https://github.com/salesforcecli/plugin-release-management/compare/v1.13.3...v1.13.4) (2021-06-25)
+
+### Bug Fixes
+
+- some of our projects use GH_TOKEN, some GITHUB_TOKEN ([55b6b83](https://github.com/salesforcecli/plugin-release-management/commit/55b6b83742804e42e1913898004e9d7f00f477d8))
+
+### [1.13.3](https://github.com/salesforcecli/plugin-release-management/compare/v1.13.2...v1.13.3) (2021-06-25)
+
+### Bug Fixes
+
+- add httpOptions agent entry for s3 ([#214](https://github.com/salesforcecli/plugin-release-management/issues/214)) ([e40153e](https://github.com/salesforcecli/plugin-release-management/commit/e40153eb1d02201a979c089e2c25675d72c77082))
+- try to get proxy to s3 to work ([#210](https://github.com/salesforcecli/plugin-release-management/issues/210)) ([2f83ada](https://github.com/salesforcecli/plugin-release-management/commit/2f83adac64cb450fbda6e770c148d4a215a37c3b))
+
 ### [1.13.2](https://github.com/salesforcecli/plugin-release-management/compare/v1.13.1...v1.13.2) (2021-06-22)
 
 ### Bug Fixes